Clean up CaffeineFragment subclasses (#5)
* Clean up CaffeineFragment subclasses

- https://app.asana.com/0/889120272558276/909108826294811
- The task was originally to address memory leaks reported by
leakcanary. We can get rid of the leak on the binding variable by
calling `binding.unbind()` in `onDestroyView()`. However, that only
touched the surface of the problem. The root `ViewGroup` of the layout
is still leaked. Here are some discussions. The issue will be fixed in
the AOSP framework, but we won't know when there will be an update from
Google.
- https://github.com/square/leakcanary/issues/1081
- https://github.com/square/leakcanary/issues/1137
- After the clean-up, the binding variable is private.
- Remove `super.onViewCreated()` to remove inconsistency across these
fragments. We don't have any plan to override this method in the base
class `CaffeineFragment` yet.
- Group @Inject member variables.
- Add the entire /.idea/ folder to gitignore.