[AND-194] Add unit tests for the lobby section headers (#399)
* [AND-194] Add unit tests for the lobby section headers

The lobby section’s name is guaranteed non-null in Lobby API V3.
We have switched to V4 where the section name can be null, which indicates
that this section shouldn’t be shown. This adds variance to the code that is no
longer guaranteed to be correct at compile time. We should add unit tests to
make sure the following scenarios are under coverage:

1. No section header from the API
2. Do not show an empty section header
3. Show a non-empty section header