Revert "[AND-394] Close the channel when the Apollo subscription call fails (#836)" (#850)
* Revert "[AND-394] Close the channel when the Apollo subscription call fails (#836)"

This reverts commit ce354ccfb069f38dd89449a01c6b0daf500e3944.

The previous fix didn't work, and it made it worse because the channel
is closed whenever there is a failure. If a user switches from Wifi to
Cellular, it will crash. We will continue to investigate the issue.

After this commit is reverted, it won't crash but the current video
won't resume. For some reason the next video will. That's probably why
some users can hear the audio from the next stage.